requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
pydantic==2.12.5
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0